Fix link issue with MinGW-W64 g++ 8.1.0 - migrate the content of vpMbtTukeyEstimator.cpp in vpMbtTukeyEstimator.h - The code previously in vpMbtTuckeyEstimator.cpp produced   a link issue with MinGW-W64 x86_64-8.1.0-posix-seh-rt_v6-rev0 (g++   8.1.0):   testTukeyEstimator.cpp:(.text.startup+0x522): undefined reference to   `vpMbtTukeyEstimator<double>::MEstimator(std::vector<double,   std::allocator<double> > const&, std::vector<double,   std::allocator<double> >&, double)'
  Note that with previous MinGW-W64 version
  x86_64-7.3.0-posix-seh-rt_v6-rev0 (g++ 7.3.0) build succeed.

  To remove this link issue the solution was to move the content of
  vpMbtTuckeyEstimator.cpp before remove.
